Move signature species to separate file
